update relations on passport database

See merge request mark-heramis/laravel-restful-api!52

+++ database/migrations/2014_07_02_230147_migration_cartalyst_sentinel.php
@@ -32,8 +32,8 @@ class MigrationCartalystSentinel extends Migration
     public function up()
     {
         Schema::create('activations', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->string('code');
             $table->boolean('completed')->default(0);
             $table->timestamp('completed_at')->nullable();
@@ -43,8 +43,8 @@ class MigrationCartalystSentinel extends Migration
         });
 
         Schema::create('persistences', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->string('code');
             $table->timestamps();
 
@@ -53,8 +53,8 @@ class MigrationCartalystSentinel extends Migration
         });
 
         Schema::create('reminders', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
             $table->string('code');
             $table->boolean('completed')->default(0);
             $table->timestamp('completed_at')->nullable();
@@ -64,7 +64,7 @@ class MigrationCartalystSentinel extends Migration
         });
 
         Schema::create('roles', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('slug');
             $table->string('name');
             $table->text('permissions')->nullable();
@@ -75,17 +75,17 @@ class MigrationCartalystSentinel extends Migration
         });
 
         Schema::create('role_users', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('role_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
             $table->nullableTimestamps();
 
             $table->engine = 'InnoDB';
         });
 
         Schema::create('throttle', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->nullable();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->nullable();
             $table->string('type');
             $table->string('ip')->nullable();
             $table->timestamps();
@@ -95,7 +95,7 @@ class MigrationCartalystSentinel extends Migration
         });
 
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('email');
             $table->string('password');
             $table->text('permissions')->nullable();
